base: add time
The package time provides a small program, from GNU project, for
measuring CPU resource usage. This is an alternative for the bash
function with the same name, when this function cannot be used.

Previously, this package was installed by recursive dependency of
lsb-core. It has been removed when we dropped lsb-core for lsb-release.

This commit actually fixes this regression.